Enable C++ 14. Fix all style and syntax warnings that followed.

+++ Parser/CompletedClockLine.cpp
@@ -47,7 +47,7 @@ QDateTime CompletedClockLine::endTime() const
     return d->end_;
 }
 
-int CompletedClockLine::duration() const
+long CompletedClockLine::duration() const
 {
     return startTime().secsTo(d->end_);
 }
